SIR-6 Refactored SirLangAssembler class
