--- src/main/java/com/sirlang/assembler/CommandMapper.java
@@ -1,4 +1,0 @@
-package com.sirlang.assembler;
-
-interface CommandMapper {
-}
