--- src/main/java/com/sirlang/assembler/CommandMapperImpl.java
@@ -1,5 +1,0 @@
-package com.sirlang.assembler;
-
-public class CommandMapperImpl implements CommandMapper {
-
-}
